backend/requirements.txt
Django
djangorestframework
web3
python-dotenv
